Fix signedness warning in tests/unicode-collate.c
tests/unicode-collate.c: In function ‘main’:
tests/unicode-collate.c:77:11: error: comparison of integer expressions of different signedness: ‘int’ and ‘guint’ {aka ‘unsigned int’} [-Werror=sign-compare]
  if (argc > i)
           ^

+++ tests/unicode-collate.c
@@ -74,7 +74,7 @@ int main (int argc, char **argv)
         }
     }
 
- if (argc > i)
+  if ((guint) argc > i)
     {
       in = g_io_channel_new_file (argv[i], "r", &error);
       if (!in)
